backend/requirements.txt
# --- Framework Web & Server ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6  # Per upload file se necessario

# --- Task Queue & Async ---
celery[redis]==5.3.6
redis==5.0.1

# --- Database & ORM ---
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # Driver PostgreSQL
alembic==1.13.1         # Per le migrazioni del DB (fondamentale in prod)

# --- HTTP Client & Scraping ---
requests==2.31.0
beautifulsoup4==4.12.3  # Per fallback su HTML parsing
httpx==0.26.0           # Client HTTP asincrono (ottimo per FastAPI)

# --- AI & Validation ---
pydantic==2.8.2
pydantic-settings==2.1.0 # Gestione .env robusta
# instructor==0.5.2      # Opzionale: Se vuoi validazione LLM avanzata subito
# mistralai==0.1.3       # Opzionale: SDK ufficiale Mistral
litellm==1.10.1  
instructor==1.3.0
google-generativeai==0.7.2

# --- Utilities ---
python-dotenv==1.0.1
tenacity==8.2.3          # Per gestire i retry automatici (es. API rate limits)

# --- FIX SCRAPING ---
feedparser==6.0.11
fake-useragent==1.5.1